transfer configuration will update the source directory to remove it when configs are transferred up
